Finished refactoring
Start writing doxygen comments to public interface

+++ CryptoLib/crypto.c
@@ -8,7 +8,7 @@
 #include "des.h"
 #include "paddings.h"
 
-int AddPadding(__in const void* input, __in uint64_t inputSize, __in PaddingType padding, __in uint64_t blockSize, __out void* output, __inout uint64_t* outputSize, __in bool fillAllBlock)
+int AddPadding(__in const void* input, __in uint64_t inputSize, __in PaddingType padding, __in size_t blockSize, __out void* output, __inout uint64_t* outputSize, __in bool fillAllBlock)
 {
     int status = NO_ERROR;
     if (status = CheckPaddingInputOutput(input, inputSize, blockSize, output, outputSize))
